Tinggal PDF dan detail saja

+++ application/views/admin/hak_cipta/index.php
@@ -31,6 +31,7 @@
                                 </div>
                             </div>
                         </div>
+                        <?php echo $this->session->flashdata('pesan');?>
                         <div class="body">
                             <div class="table-responsive">
                                 <table class="table table-bordered table-striped table-hover js-basic-example dataTable">
@@ -70,7 +71,7 @@
                                             <td><?=$cipta->status_trs?></td>
                                             <td style="width: 70px;">
                                                 <a class="btn btn-primary btn-sm px-2" href="<?=base_url()?>admin/hak_cipta/edit/<?=$cipta->id?>"><i class="zmdi zmdi-edit"></i></a>
-                                                <a class="btn btn-danger btn-sm" href="<?=base_url()?>admin/hak_cipta/delete/<?=$cipta->id?>"><i class="zmdi zmdi-delete"></i></a>
+                                                <a onclick="return confirm('Apakah anda yakin untuk menghapus data ini (<?=$cipta->judul?>)?');" class="btn btn-danger btn-sm" href="<?=base_url()?>admin/hak_cipta/delete/<?=$cipta->id?>"><i class="zmdi zmdi-delete"></i></a>
                                             </td>
                                         </tr>
                                     <?php } ?>
@@ -239,7 +240,7 @@
                         }).change();
                     </script>
 
-                    <small>File <strong>Scan Contoh Karya/Ciptaan</strong> yang diunggah haruslah kurang dari 20 MB, format file (pdf/jpg)</small>
+                    <small>File <strong>Scan Contoh Karya/Ciptaan</strong> yang diunggah haruslah kurang dari 20 MB</small>
                     <div class="custom-file mb-2">
                         <input type="file" class="custom-file-input" name="file_ciptaan" id="file_ciptaan">
                         <label class="custom-file-label" for="file_ciptaan">Scan Contoh Karya/Ciptaan</label>
